# WarehouseApi/Entities/Category.cs
using System.ComponentModel.DataAnnotations;

namespace WarehouseApi.Entities
{
    public class Category
    {
        public int Id{ get; set; }
        public string Name { get; set; } = string.Empty;
        
        public string Description { get; set; } = string.Empty;

        public ICollection<Product> Products { get; } = new List<Product>();

    }
}